Use fixed-size types where appropriate
-Replace many uses of "long" with "uint32_t"

-Replace some uses of "short" with "uint16_t"

-Replace some uses of "unsigned char" with
 "uint8_t"

-Minor Makefile improvements included as well

+++ bin.c
@@ -36,7 +36,7 @@ do_cry(unsigned char red, unsigned char green, unsigned char blue)
 int
 main()
 {
-	long i;
+	uint32_t i;
 	unsigned char red, green, blue;
 	unsigned int output;
 	FILE *f;
